adaption for obcloud support ssl connect
