+++ server/plugins/connect-plugin-ob-mysql/src/main/java/com/oceanbase/odc/plugin/connect/obmysql/OBMySQLConnectionExtension.java
@@ -156,11 +156,10 @@ public class OBMySQLConnectionExtension implements ConnectionExtensionPoint {
         } catch (Exception e) {
             Throwable rootCause = ExceptionUtils.getRootCause(e);
             if (rootCause == null) {
-                log.warn("Failed to get connection, errMsg={}", ExceptionUtils.getRootCauseReason(e));
+                log.warn("Failed to get connection when test connection, jdbcUrl={} ", jdbcUrl, e);
                 return TestResult.unknownError(e);
             }
-            log.warn("Failed to get connection, rooCauseErrMsg={}, errMsg={}", rootCause.getLocalizedMessage(),
-                    ExceptionUtils.getRootCauseReason(e));
+            log.warn("Failed to get connection when test connection, jdbcUrl={} ", jdbcUrl, e);
             String host = hostAddress.getHost();
             Integer port = hostAddress.getPort();
             if (rootCause instanceof ConnectException) {
